separate "Electornic life" project into three modules
