+++ 01-syntax/09-modules/electronic-life/index.html
@@ -4,6 +4,5 @@
 <body>
 <script src="moduleSupport.js"></script>
 <script src="animateworld.js"></script>
-<script src="elife-all.js"></script>
 <script>animateWorld(valley);</script>
 </body>
